First passing unit-test. The idea is that systemsetup will return a status code for each component successfully activated by reading the status of the special function registe. There were (at least) two possible approaches. 1. To do it like now where you read the sfr with IO_Read apply the bit operation with IO_Write and read again the SFR with a view to generate a status flag that is going to be returned and then checked in the unit test. This feels more natural as the mock expectations match exactly how we do it in the production code. 2. This approach is to do an IO_Read inside the test fixture right in the check function which is not very intuitive as we will have to add a mock expectation for the final IO_Read which is not going to be reflected in the production code. Another approach would be to make sure that during IO_Write we also have a line *address = value. However it feels like it is kind of cheating.

+++ include/io/IO.h
@@ -31,7 +31,7 @@
 #include <stdint.h>
 
 typedef uint32_t ioAddress;
-typedef uint16_t ioData;
+typedef uint32_t ioData;
 
 ioData IO_Read(ioAddress *address);
 void  IO_Write(ioAddress *address, ioData data);
